Updates speed control to use float

+++ modules/3D.js
@@ -22,7 +22,7 @@ class iiiD extends modV.Module3D {
 		this.add(new modV.RangeControl({
 			variable: 'speed',
 			label: 'Speed',
-			varType: 'int',
+			varType: 'float',
 			min: 0,
 			max: 20,
 			step: 1,
